fix type error in toFloat

+++ ThrEd4/wrappers.h
@@ -247,10 +247,10 @@ template <class inType> auto toDouble(inType invar) noexcept -> double {
 }
 
 template <class inType>
-auto toFloat(inType invar) noexcept(!(std::is_same_v<inType, float> ||
+auto toFloat(inType invar) noexcept(!(std::is_same_v<inType, double> ||
                                       (std::is_same_v<inType, size_t> && !std::is_same_v<uint32_t, size_t>)))
     -> float {
-  if constexpr (std::is_same_v<inType, float>) {
+  if constexpr (std::is_same_v<inType, double>) {
 	try {
 	  return gsl::narrow<float>(invar);
 	}
@@ -280,7 +280,7 @@ auto toFloat(inType invar) noexcept(!(std::is_same_v<inType, float> ||
 template <class inType>
 auto toSize(inType invar) noexcept(!(std::is_signed_v<inType> || (std::is_same_v<inType, uint64_t> && std::is_same_v<size_t, uint32_t>)))
     -> size_t {
-  static_assert((std::is_signed_v<inType> || std::is_same_v<inType, unsigned long long> ||
+  static_assert((std::is_signed_v<inType> || std::is_same_v<inType, uint64_t> ||
                  std::is_same_v<inType, uint8_t> || std::is_same_v<inType, uint16_t> ||
                  std::is_same_v<inType, uint32_t> || std::is_same_v<inType, DWORD>),
                 "cannot use wrap::toSize here.");
